app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
apispec==6.7.0
apispec-marshmallow-advanced==0.4
apispec-webframeworks==1.2.0
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
caio==0.9.17
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-JSONRPC==3.0.1
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
logging==0.4.9.6
logging_tree==1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pillow==11.2.1
pluggy==1.5.0
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.9.1
pydantic_core==2.27.1
Pygments==2.19.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
Python_IO==0.3
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tornado==6.4.1
typeguard==2.13.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.17.2
